mcp_github_agent/requirements.txt
azure-ai-projects
azure-ai-agents
azure-identity
python-dotenv
